add lib/capabilities.d.ts
